+++ src/autogenerated/ResourcesAutogen.hpp
@@ -26,8 +26,6 @@ public:
         QPixmap banRed;
         QPixmap copyDark;
         QPixmap copyLight;
-        QPixmap loginCopyDark;
-        QPixmap loginCopyLight;
         QPixmap menuDark;
         QPixmap menuLight;
         QPixmap mod;