+++ src/singletons/Theme.cpp
@@ -88,12 +88,10 @@ void Theme::actuallyUpdate(double hue, double multiplier)
     if (this->isLightTheme())
     {
         this->buttons.copy = getResources().buttons.copyDark;
-        this->buttons.loginCopy = getResources().buttons.loginCopyDark;
     }
     else
     {
         this->buttons.copy = getResources().buttons.copyLight;
-        this->buttons.loginCopy = getResources().buttons.loginCopyLight;
     }
 }