+++ src/singletons/Theme.hpp
@@ -50,7 +50,6 @@ public:
 
     struct {
         QPixmap copy;
-        QPixmap loginCopy;
     } buttons;
 
     void normalizeColor(QColor &color);